--- a/tensorflow/lite/experimental/litert/vendors/qualcomm/compiler/graph_mapper.h
+++ b/tensorflow/lite/experimental/litert/vendors/qualcomm/compiler/graph_mapper.h
@@ -86,9 +86,6 @@ class GraphMapper {
   // Finalize QNN Graph. Call this after all ops have been mapped.
   LiteRtStatus Finalize();
 
-  // Pick graph config based on subgraph.
-  absl::Span<const QnnGraph_Config_t*> PickGraphConfigHeuristic();
-
   inline void RegisterOutput(LiteRtTensor litert_tensor) {
     graph_outpus_.insert(litert_tensor);
   }