1. Use O3 as default
2. Use relax precision for float as default because QNN tends to always enable it.

# Test
```
     0.0ms [  INFO ]  <I> QnnContext_getBinary done successfully. context = 0x1
INFO: [tensorflow/lite/experimental/litert/vendors/qualcomm/qnn_manager.cc:247] Serialized a context bin of size (bytes): 644400

INFO: [tensorflow/lite/experimental/litert/vendors/qualcomm/compiler/qnn_compiler_plugin.cc:437] Context binary 0 generated
     0.0ms [  INFO ]  <I> QnnContext_free started. context = 0x1
     0.0ms [  INFO ]  <I> QnnContext_free done successfully.
     0.0ms [  INFO ]  <I> QnnDevice_free started. device = 0x1
     0.0ms [  INFO ]  <I> QnnDevice_free done. status 0x0
     0.0ms [  INFO ]  <I> QnnBackend_free started. backend = 0x1
     0.0ms [  INFO ] Closing environment: 0x55aab477e5f0
[       OK ] SupportedOpsTest/QnnPluginOpCompatibilityTest.SupportedOpsTest/41 (666 ms)
[----------] 42 tests from SupportedOpsTest/QnnPluginOpCompatibilityTest (2995 ms total)

[----------] Global test environment tear-down
[==========] 97 tests from 5 test suites ran. (3196 ms total)
[  PASSED  ] 97 tests.
```